requirements.txt
# =============================================================================
# Azure Agentic Engineering - Python Requirements
# =============================================================================
# Install with: pip install -r requirements.txt

# -----------------------------------------------------------------------------
# Azure AI SDKs
# -----------------------------------------------------------------------------
azure-identity>=1.15.0
azure-ai-projects>=1.0.0b1
azure-ai-inference>=1.0.0b1
azure-ai-evaluation>=1.0.0b1
azure-search-documents>=11.4.0

# Microsoft Agent Framework (preview)
agent-framework --pre

# OpenAI SDK
openai>=1.12.0

# -----------------------------------------------------------------------------
# Azure Management SDKs
# -----------------------------------------------------------------------------
azure-mgmt-resource>=23.0.0
azure-mgmt-cognitiveservices>=13.5.0
azure-keyvault-secrets>=4.7.0

# -----------------------------------------------------------------------------
# Development Tools
# -----------------------------------------------------------------------------
black>=24.0.0
isort>=5.13.0
pylint>=3.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
mypy>=1.8.0

# -----------------------------------------------------------------------------
# Web & API
# -----------------------------------------------------------------------------
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pydantic>=2.5.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0
rich>=13.7.0
structlog>=24.1.0
tenacity>=8.2.0

# -----------------------------------------------------------------------------
# Observability
# -----------------------------------------------------------------------------
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
azure-monitor-opentelemetry>=1.2.0

# -----------------------------------------------------------------------------
# Jupyter (optional, for notebooks)
# -----------------------------------------------------------------------------
jupyter>=1.0.0
ipykernel>=6.29.0
